Bug fixes, set checking, oversight correction
Corrected an oversight where, if a comic page has a content warning and no link leading either to the next page or to an alt page, the content warning cannot be removed. Relevant modifications to both PHP and javascript have been added to correct this issue.

Additional variable checking and setting.