Changed $bloginfo['comicname'] to $ccpage->title
"$bloginfo['comicname']" isn't anything. Making the assumption that either "$ccsite->sitetitle" or "$ccpage->title" was meant instead. Went with the latter.

In the case of "comic" modules, links in the admin bar only generate if the variables they require to work properly are set. This will likely be modified further at a later date, as "add" no longer appears when in the comic archive as a result of this change and there's no real reason for it not to that I can think of.