Significant modifications made to how content warnings work.
If the archive has comic thumbnails turned on, and a thumbnail's corresponding page has a content warning, these thumbnails are now blurred with the words "Content Warning" imposed over the top of them.

Additionally, hovering over these thumbnails reveal the content warning text in full. Thumbnail alt text also contains the content warning for those who can't read hovertext due to the use of a screenreader. If this is an incorrect use of alt text, it will be removed.

Page title text, if displayed with thumbnails, is no longer a link. This might be reverted later if it causes an issue with navigation, although at the moment there doesn't seem to be a problem. I easily could've missed something while testing, though.

Content warnings on actual pages have had their HTML markup rewritten. They are now significantly easier to style.

A "warning_toggle" cookie is now referenced. This will be properly implemented later.