Bug Fix: All $db variables input changed from doublequotes to singlequotes
It is possible for a randomly generated database password to contain a $ character. 

I can't speak for other programming languages, but in PHP, anything that looks like it could be a variable while sitting beteween doublequotes gets parsed as a variable instead of as a string, without any sort of concatenation required.

So if you generate a random password for your database while installing ComicControl, and if that password contains a $ anywhere, because the actual final output for dbconfig.php has your passwords and everything sitting between doublequotes, that $ in your password and anything valid after it would be mistaken for a variable, and this would cause your password to just not work, even though it's correct, because it's being read as a variable and not as a string.

So everything that is user supplied (not limited to $dbpass, but everything) is now sitting between escaped singlequotes, hopefully solving this issue and any potential variations of it completely.